Create chatBot
Not sure whether I should have pulled / added to the main branch so I went with this. 

If this doesn't work, please let me know and I can push this as a commit to to the main branch or something.

That being said, this is not completed. It currently goes up to (8) different routes, most of which consist of basic interaction with the user and the path once a user inputs the keywords 'forgot' and 'password'

Created using swing and consequent JFRAME classes for the user interface and application
Created using: Java-Eclipse